add reset and update password endpoints
to add the endpoints, using templates, i had to use fastapi's forms, and the constraint is that all the variables in the form are passed as parameters in the corresponding function. Also, unable to get the form data from request: Request (fastapi), so had to fallback to the above method. Also, wasn't able to authenticate from via token (how to do it? send the token as some header in the template?)

the flow is: if user forgets the password, user sends a post request to /request-reset-password, which generates a token, and sends the corresponding token to the email given, (only if there is a user with that email).

then, the user clicks on the link in the email, which makes a get request to /reset-password with token as a query parameter, this link first verifies the token, then renders the template for filling the new password (it sends the token as a hidden input too), which is submitted to /reset-password via a post request.

which checks if the new passwords are matching, and if they are, is the token valid, and if it is, then changes the password in the db.

a similar flow happens for update_password, where a logged in user makes this request, and sends the access token as a query parameter, and gets a form in return, which submits to /update-password via post method, and send the old_password, new_password, confirm_new_password and token, and using them, the password is updated in the db.

also, added endpoint for getting the new access token, using the refresh token provided at the time of login.

updated mongo_service to work according to the above implementation.